Cambio en formas de pedir datos y uso de floats en 2 variables

+++ escape/escape.rb
@@ -1,6 +1,6 @@
 g, r = ARGV
 
-result = 2 * ARGV[0].to_f * ARGV[1].to_i
+result = 2 * ARGV[0].to_f * ARGV[1].to_f
 
 escape = Math.sqrt(result)
